Add dev auth_token to Helm chart
The auth_token was implemented for authenticating the UI, and
the dev value was correctly set through the environment via
docker-compose. However the Kind cluster setup uses a version
of pctasks-server deployed via the Helm chart, so that setting
needs to be set through the Helm chart for integration tests to work.
This commit adds the dev auth token to the Helm chart and sets
it for the Kind cluster dev deployment. Without this change,
the server was crashing on deployment due to the lack of
a required setting, causing submitted workflows to fail
even if they were submitted with the correct API key.