Ensure 1 node for test
This configures a minimum of one low priority node for pctasks-test, to
speed up the feedback loop when developing / deploying a pipeline.

The `min_low_priority` terraform variable was used for two pools,
"tasks" and "ingest". I've split that into two variables and left ingest
at the previous default of 0.