Fix issue with local dir sorting in os.walk
Strangely, the behavior of os.walk is different for
the dev container running on my local machine (Windows
11 w/ WSL 2 running Ubuntu 20.04) vs. the GitHub Actions
runner (which I believe is running Ubuntu 20.04).  The
ordering of os.walk, while top-down in both cases, is
doing depth-first in the GitHub Actions case and breadth-
first in the local case.  The LocalStorage code was relying
on the breadth-first behavior, and this change makes it
work in both cases.