Fetch PC API tokens in executor, not locally.
This change defers the evaluation of the
${{ pc.get_token() }} expression to the executor, rather than
evaluating it locally. Before this change,
workflows written to the would have short-expiry tokens
embedded into them; now, the token is generated as the task
is being prepared for submission, and so will have a longer
relevant lifetime.

Also, renamed task_config to definition in the task run
messages