[dataset]: Add flag to generate update workflows
This is a bit of a hack, but it's a relatively small change to generate
an "update" workflow from a dataset.yaml.

When you pass `--is-update`, this will

1. Append a `--since` argument to the workflow arguments
2. Inject the string literal `"${{ args.since }}"` for the `since`
   argument to the create-splits task. Then, when the update is later
   run with a concrete `--since` that concrete value will be templated
   in.

Combine that with a chunkset ID that also contains the literal
`${{ args.since }}` and I think we have everything we need for dynamic
updates.