[CI]: Mount the azure directory in deploy (#137)
This mounts ~/.azure into the docker container doing the deployment. We
now get credentials from the Azure CLI, which are stored in ~/.azure on
the GitHub runner. By mounting the directory to `/root/.azure` on the
docker container deploying things, it too has access to the Azure
credentials from the azlogin action.