Validate check for failed batch task
This fixes a bug where the job partition check for
failed batch tasks was making a check against
tasks that were not part of that job partition.
This would cause any failed batch task to be
fetched from "indexed_task_ids", even if
it wasn't part of the current job partition.
This scopes the check so that only tasks that
are part of the current job partition are
checked.