Consolidate comsosdb connections
This changes the way we connect to cosmosdb. Instead of creating a new
connection for each thread, we create a single connection per container
and model and pass it to threads. This is to combat some issues we've
seen where the client creation times out on making a connection to
cosmosdb; the documentation suggests that clients should be kept
to a minimum.

Tested on a job with about 50 partitions, and it seems to work fine.
Also, this issue seems to indicate that the Azure SDK is thread-safe:
https://github.com/Azure/azure-sdk-for-python/issues/28665