[core]: Avoid PyZipFile in code upload
This reimplements code upload for directories to avoid
PyZipFile.writepy. It (deliberately) uploads `.pyc` files, which imposes
a strong constraint that the version of Python running submit on the client
matches the version of Python running in the task.

Instead, we replicate PyZipFile.writepy's behavior, but we upload `.py`
files rather than `.pyc`. The zip-based importer works with both `.py`
and `.pyc` files.