Set PCTASKS_COSMOSDB__TEST_CONTAINER_SUFFIX in Functions
This sets the `PCTASKS_COSMOSDB__TEST_CONTAINER_SUFFIX` environment
variable in the Azure Functions, to ensure that the test Cosmos DB
containers are used.

We're using Azure Function's "Binding expression patterns"
(https://learn.microsoft.com/en-us/azure/azure-functions/functions-bindings-expressions-patterns)
to dynamically set the Cosmos DB container to watch. This required
updating the terraform to set the container name to watch as an app
setting.

This did require a small change to the names of the temporary Cosmos DB
containers we create for testing. Previously they were
`tmp-{name}-{suffix}-{uuid}`.

To make templating possible in the docker-compose file, I went with
`{name}{suffix}`, where suffix is uniquely set to the GitHub actions run
ID, so should be sufficient for avoiding clashes.