Expand retried exceptions in with_backoff
Adds `TimeoutError` and `ConnectionError` to the list of retried
exceptions in `with_backoff`.

It's not clear to me whether or not this was previously being retried,
but the goes-glm pipeline has failed with

```
requests.exceptions.ConnectionError:
HTTPSConnectionPool(host='goeseuwest.blob.core.windows.net', port=443):
Read timed out.
```

I'd have thought that `azure-storage-blob` is retrying that, but I'm not
sure.