Optimize task images
This PR optimzies the various task images (task-base, run, and ingest).
It removes *a bunch* of stuff, some of which is probably needed for at
least some of our pipelines.

I'm going to do an audit of which pipelines are using which images. I'm
also a bit unsure about what sorts of issues are going to be picked up
by CI. We might need additional testing to be confident that this
doesn't break anything.

Here's the size comparison:

| image             | before | after |
| ----------------- | ------ | ----- |
| pctasks-ingest    | 1.3GB  | 257MB |
| pctasks-run       | 630MB  | 290MB |
| pctasks-task-base | 3.32   | 373MB |