[low-latency]: Added goes-cmi (#247)
This adds a streaming workflow for goes-cmi. It's running in the
staging-test deployment:

```
» curl -s
"https://pct-apis-staging.westeurope.cloudapp.azure.com/stac/collections/goes-cmi/items?limit=1"
| jq -r .features[0].properties.datetime
2023-07-25T20:23:55.4Z
```

While tuning the CPU and memory requests for this task, I noticed that
we were using a different VM size on the streaming tasks than what we're
using on the batch tasks. I think those should match, so I updated
streaming to be `D3_v2`.