[low-latency]: Update storage event dispatch
Previously, storage event dispatch was handled by regular expressions in
the code.

This moves the dispatch to a configuraiton system. See the docstring in
`StorageEventsCF/__init__.py` for an explaination, but the basic idea is
to use a set of environment variables to define the queue name, prefix
filter, and suffix filter to use for matching storage events to queues.