Structured logs for low-latency functions (#255)
* Structured logs for low-latency functions

This adds some more structured logging to the Azure Functions supporting
low-latency pipelines. I intend to build some monitoring and dashboards
on top of these messages.

Unfortunately, the Azure Functions runtime overwrites any properties we
put in `{extra: {custom_dimensions: ...}}`. So as a workaround we'll
put the structured message in the `message` field as JSON and parse
that in the queries.