[ENH]: Allow for matching on the full path
This adds a parameter to `Storage.walk`, allowing users to specif
whether the `matches` applies just to the filename or to the full
path, including the filename.

The same keyword is added to the task and task input, to let this be
set in the datasset.yaml.

Recent changes to the data published in the ecmwf container requires a
new feature in pctasks' create-chunks: the ability to match on a full
path rather than just a filename.

Data are being published under both the

    - <date-prefix>/ifs/...
    - <date-prefix>/aifs/...

We list both sets of data, but our stactools package can't handle the
aifs data properly yet. So we want to filter it out.

Previously, `matches` could only target the filename being listed.
We need that to filter to specific products. We want to extend that
match expression to include the prefix, so that we can filter out the
`aifs` data.